backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
flask-cors
python-dotenv
werkzeug
resend